Escape right square brackets
As discussed in #148, right square brackets need to be escaped for
correct handling in links in some cases.  This is an alternative to
that PR that is based on current sources and adds the requested
testcase, since there was no response to the request in that PR for a
testcase to be added to the testsuite.