Avoid stripping nonbreaking spaces
Nonbreaking spaces should be preserved in places such as the start of
a paragraph or blockquote, so change various places to strip only
ASCII `' \t\r\n'`.  There may be other places that should also avoid
stripping nonbreaking spaces (or, conversely, where *trailing* such
spaces could safely be stripped even if they no longer are after this
change), but this seems a reasonable starting point to fix issues in
this area.